refactor(wallpaper): split monolithic module into focused components
- Extract type definitions to dedicated types.ts file
- Move search functionality to search.ts module
- Separate wallpaper setting logic into set.ts
- Create thumbnail handling module (thumbnail.ts)
- Update all import paths to reference new modular structure
- Maintain existing API surface while improving code organization

This improves maintainability by separating concerns and making the
codebase easier to navigate and modify.
