+++ src/electron/main/trpc/routes/api/unsplash/index.ts
@@ -1,7 +1,7 @@
 import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 import { publicProcedure, router } from "@electron/main/trpc/index.js";
-import { type ApiWallpaper } from "@electron/main/trpc/routes/wallpaper/index.js";
+import { type ApiWallpaper } from "@electron/main/trpc/routes/wallpaper/types.js";
 
 interface UnsplashPhoto {
   id: string;
